added error handling for non existing folders

+++ src/app/__init__.py
@@ -99,17 +99,24 @@ from .commands import handle_command
 from .sockets import emit_status  # TODO: dafuq is this, send help
 from .import routes
 
-# cleaning upload folder
-SHL.output(f"{green2}Cleaning Upload folder.{white}", "Upload")
-for the_file in os.listdir(app.config['UPLOAD_FOLDER']):
-    file_path = os.path.join(app.config['UPLOAD_FOLDER'], the_file)
-    try:
-        if os.path.isfile(file_path):
-            os.unlink(file_path)
-        elif os.path.isdir(file_path):
-            shutil.rmtree(file_path)
-    except Exception as e:
-        SHL.output(f"{red} ERROR: Cleaning Upload folder., Reason: {e}{white}", "Upload")
+
+# checking and creating upload dir
+if not os.path.exists(app.config['UPLOAD_FOLDER']):
+    os.makedirs(app.config['UPLOAD_FOLDER'], exist_ok=True)
+    SHL.output(f"{green2}Upload folder was not present, created upload folder.{white}", "Upload")
+
+else:
+    # cleaning upload folder
+    SHL.output(f"{green2}Cleaning Upload folder.{white}", "Upload")
+    for the_file in os.listdir(app.config['UPLOAD_FOLDER']):
+        file_path = os.path.join(app.config['UPLOAD_FOLDER'], the_file)
+        try:
+            if os.path.isfile(file_path):
+                os.unlink(file_path)
+            elif os.path.isdir(file_path):
+                shutil.rmtree(file_path)
+        except Exception as e:
+            SHL.output(f"{red} ERROR: Cleaning Upload folder., Reason: {e}{white}", "Upload")
 
 
 # I left this for testing
